+++ crates/hale-codegen/src/codegen.rs
@@ -30,6 +30,7 @@ use crate::stdlib::env::EnvStdlib;
 use crate::stdlib::io_file::IoFileStdlib;
 use crate::stdlib::io_fs::IoFsStdlib;
 use crate::stdlib::io_stdin::IoStdinStdlib;
+use crate::stdlib::io_tls::IoTlsStdlib;
 use crate::stdlib::math::MathStdlib;
 use crate::stdlib::process::ProcessStdlib;
 use crate::stdlib::rand::RandStdlib;
@@ -29870,184 +29871,8 @@ impl<'ctx, 'p> Cx<'ctx, 'p> {
         Ok((ret_i64.into(), CodegenTy::Int))
     }
 
-    /// `std::io::tls::connect(host, port) -> Int fallible(IoError)`.
-    /// Same fallible shape as `tcp::connect` — the C primitive
-    /// opens a TCP socket, wraps in SSL, performs the TLS
-    /// handshake, returns an opaque handle (>=0) or -1 on error.
-    fn lower_std_io_tls_connect_fallible(
-        &mut self,
-        args: &[Expr],
-        scope: &Scope<'ctx>,
-    ) -> Result<FallibleCallResult<'ctx>, CodegenError> {
-        if args.len() != 2 {
-            return Err(CodegenError::Unsupported(format!(
-                "std::io::tls::connect takes 2 args (host, port), got {}",
-                args.len()
-            )));
-        }
-        let (host_val, host_ty) = self.lower_expr(&args[0], scope)?;
-        if !matches!(host_ty, CodegenTy::String | CodegenTy::StringView) {
-            return Err(CodegenError::Unsupported(format!(
-                "std::io::tls::connect: host must be String, got {:?}",
-                host_ty
-            )));
-        }
-        let host_val = self.unpack_view_if_needed(host_val, &host_ty)?;
-        let (port_val, port_ty) = self.lower_expr(&args[1], scope)?;
-        if port_ty != CodegenTy::Int {
-            return Err(CodegenError::Unsupported(format!(
-                "std::io::tls::connect: port must be Int, got {:?}",
-                port_ty
-            )));
-        }
-        let i16_t = self.context.i16_type();
-        let i32_t = self.context.i32_type();
-        let i64_t = self.context.i64_type();
-        let port_i16 = self
-            .builder
-            .build_int_truncate(port_val.into_int_value(), i16_t, "tls.port.i16")
-            .map_err(|e| CodegenError::LlvmEmit(e.to_string()))?;
-        let f = self
-            .module
-            .get_function("lotus_tls_connect")
-            .expect("lotus_tls_connect declared");
-        let h_i32 = self
-            .builder
-            .build_call(
-                f,
-                &[host_val.into(), port_i16.into()],
-                "tls.connect.handle",
-            )
-            .map_err(|e| CodegenError::LlvmEmit(e.to_string()))?
-            .try_as_basic_value()
-            .left()
-            .expect("returns i32")
-            .into_int_value();
-        let is_err = self
-            .builder
-            .build_int_compare(
-                inkwell::IntPredicate::SLT,
-                h_i32,
-                i32_t.const_zero(),
-                "tls.connect.is_err",
-            )
-            .map_err(|e| CodegenError::LlvmEmit(e.to_string()))?;
-        let h_i64 = self
-            .builder
-            .build_int_s_extend(h_i32, i64_t, "tls.connect.handle.i64")
-            .map_err(|e| CodegenError::LlvmEmit(e.to_string()))?;
-        self.complete_io_fallible_call(
-            is_err,
-            host_val,
-            Some((h_i64.into(), CodegenTy::Int)),
-            "tls.connect",
-        )
-    }
 
-    /// `std::io::tls::send_bytes(handle: Int, bytes: Bytes) -> Int`.
-    /// Non-fallible at the language level: returns 0 on success or
-    /// -1 on error. Mirrors tcp::__send_bytes.
-    fn lower_std_io_tls_send_bytes(
-        &mut self,
-        args: &[Expr],
-        scope: &Scope<'ctx>,
-    ) -> Result<(BasicValueEnum<'ctx>, CodegenTy), CodegenError> {
-        if args.len() != 2 {
-            return Err(CodegenError::Unsupported(format!(
-                "std::io::tls::send_bytes takes 2 args (handle, bytes), got {}",
-                args.len()
-            )));
-        }
-        let (h_val, h_ty) = self.lower_expr(&args[0], scope)?;
-        if h_ty != CodegenTy::Int {
-            return Err(CodegenError::Unsupported(format!(
-                "std::io::tls::send_bytes: handle must be Int, got {:?}",
-                h_ty
-            )));
-        }
-        let (b_val, b_ty) = self.lower_expr(&args[1], scope)?;
-        if !matches!(b_ty, CodegenTy::Bytes | CodegenTy::BytesView) {
-            return Err(CodegenError::Unsupported(format!(
-                "std::io::tls::send_bytes: bytes must be Bytes, got {:?}",
-                b_ty
-            )));
-        }
-        let b_val = self.unpack_view_if_needed(b_val, &b_ty)?;
-        let i32_t = self.context.i32_type();
-        let i64_t = self.context.i64_type();
-        let h_i32 = self
-            .builder
-            .build_int_truncate(h_val.into_int_value(), i32_t, "tls.sb.h.i32")
-            .map_err(|e| CodegenError::LlvmEmit(e.to_string()))?;
-        let f = self
-            .module
-            .get_function("lotus_tls_send_bytes")
-            .expect("lotus_tls_send_bytes declared");
-        let ret_i32 = self
-            .builder
-            .build_call(f, &[h_i32.into(), b_val.into()], "tls.send_bytes.ret")
-            .map_err(|e| CodegenError::LlvmEmit(e.to_string()))?
-            .try_as_basic_value()
-            .left()
-            .expect("returns i32")
-            .into_int_value();
-        let ret_i64 = self
-            .builder
-            .build_int_s_extend(ret_i32, i64_t, "tls.send_bytes.i64")
-            .map_err(|e| CodegenError::LlvmEmit(e.to_string()))?;
-        Ok((ret_i64.into(), CodegenTy::Int))
-    }
 
-    /// `std::io::tls::recv_bytes(handle: Int, max: Int) -> Bytes`.
-    /// Non-fallible: returns up to max bytes on success, or an
-    /// empty Bytes on error / peer-closed. Mirrors tcp::__recv_bytes.
-    fn lower_std_io_tls_recv_bytes(
-        &mut self,
-        args: &[Expr],
-        scope: &Scope<'ctx>,
-    ) -> Result<(BasicValueEnum<'ctx>, CodegenTy), CodegenError> {
-        if args.len() != 2 {
-            return Err(CodegenError::Unsupported(format!(
-                "std::io::tls::recv_bytes takes 2 args (handle, max), got {}",
-                args.len()
-            )));
-        }
-        let (h_val, h_ty) = self.lower_expr(&args[0], scope)?;
-        if h_ty != CodegenTy::Int {
-            return Err(CodegenError::Unsupported(format!(
-                "std::io::tls::recv_bytes: handle must be Int, got {:?}",
-                h_ty
-            )));
-        }
-        let (max_val, max_ty) = self.lower_expr(&args[1], scope)?;
-        if max_ty != CodegenTy::Int {
-            return Err(CodegenError::Unsupported(format!(
-                "std::io::tls::recv_bytes: max must be Int, got {:?}",
-                max_ty
-            )));
-        }
-        let i32_t = self.context.i32_type();
-        let h_i32 = self
-            .builder
-            .build_int_truncate(h_val.into_int_value(), i32_t, "tls.rb.h.i32")
-            .map_err(|e| CodegenError::LlvmEmit(e.to_string()))?;
-        let max_i32 = self
-            .builder
-            .build_int_truncate(max_val.into_int_value(), i32_t, "tls.rb.max.i32")
-            .map_err(|e| CodegenError::LlvmEmit(e.to_string()))?;
-        let f = self
-            .module
-            .get_function("lotus_tls_recv_bytes")
-            .expect("lotus_tls_recv_bytes declared");
-        let ptr = self
-            .builder
-            .build_call(f, &[h_i32.into(), max_i32.into()], "tls.rb.ret")
-            .map_err(|e| CodegenError::LlvmEmit(e.to_string()))?
-            .try_as_basic_value()
-            .left()
-            .expect("returns ptr");
-        Ok((ptr, CodegenTy::Bytes))
-    }
 
     /// m105: lower `std::bus::__local_dispatch(subject: String,
     /// wire_bytes: Bytes) -> ()`. Hands wire bytes (received by an
@@ -30121,51 +29946,6 @@ impl<'ctx, 'p> Cx<'ctx, 'p> {
         Ok((i64_t.const_zero().into(), CodegenTy::Int))
     }
 
-    /// `std::io::tls::close(handle: Int) -> Int`. Non-fallible:
-    /// returns 0 on success, -1 on error (bad handle). Mirrors
-    /// tcp::close_fd.
-    fn lower_std_io_tls_close(
-        &mut self,
-        args: &[Expr],
-        scope: &Scope<'ctx>,
-    ) -> Result<(BasicValueEnum<'ctx>, CodegenTy), CodegenError> {
-        if args.len() != 1 {
-            return Err(CodegenError::Unsupported(format!(
-                "std::io::tls::close takes 1 arg (handle), got {}",
-                args.len()
-            )));
-        }
-        let (h_val, h_ty) = self.lower_expr(&args[0], scope)?;
-        if h_ty != CodegenTy::Int {
-            return Err(CodegenError::Unsupported(format!(
-                "std::io::tls::close: handle must be Int, got {:?}",
-                h_ty
-            )));
-        }
-        let i32_t = self.context.i32_type();
-        let i64_t = self.context.i64_type();
-        let h_i32 = self
-            .builder
-            .build_int_truncate(h_val.into_int_value(), i32_t, "tls.close.h.i32")
-            .map_err(|e| CodegenError::LlvmEmit(e.to_string()))?;
-        let f = self
-            .module
-            .get_function("lotus_tls_close")
-            .expect("lotus_tls_close declared");
-        let ret_i32 = self
-            .builder
-            .build_call(f, &[h_i32.into()], "tls.close.ret")
-            .map_err(|e| CodegenError::LlvmEmit(e.to_string()))?
-            .try_as_basic_value()
-            .left()
-            .expect("returns i32")
-            .into_int_value();
-        let ret_i64 = self
-            .builder
-            .build_int_s_extend(ret_i32, i64_t, "tls.close.i64")
-            .map_err(|e| CodegenError::LlvmEmit(e.to_string()))?;
-        Ok((ret_i64.into(), CodegenTy::Int))
-    }
 
 
 
@@ -30361,21 +30141,6 @@ impl<'ctx, 'p> Cx<'ctx, 'p> {
         )
     }
 
-    /// Phase 1: lower `std::io::tls::recv_into(handle: Int,
-    /// buf: Bytes, max_bytes: Int) -> Int`. SSL_read into the
-    /// builder's tail. Same return semantics as tcp_recv_into.
-    fn lower_std_io_tls_recv_into(
-        &mut self,
-        args: &[Expr],
-        scope: &Scope<'ctx>,
-    ) -> Result<(BasicValueEnum<'ctx>, CodegenTy), CodegenError> {
-        self.lower_recv_into_common(
-            args,
-            scope,
-            "lotus_tls_recv_into",
-            "std::io::tls::recv_into",
-        )
-    }
 
     /// Phase 1: lower `std::io::udp::recv_into(fd: Int, buf: Bytes,
     /// max_bytes: Int) -> Int`. Single recvfrom into the builder's
@@ -30397,7 +30162,7 @@ impl<'ctx, 'p> Cx<'ctx, 'p> {
     /// Shared lowering for the recv_into family. Validates arg
     /// types, extracts the builder handle from the BytesBuilder
     /// locus, truncates fd to i32, and emits the call.
-    fn lower_recv_into_common(
+    pub(crate) fn lower_recv_into_common(
         &mut self,
         args: &[Expr],
         scope: &Scope<'ctx>,
